update and add Tests

+++ Tests/NoteDetectionTests/MIDINoteDetectorTest.swift
@@ -31,7 +31,7 @@ class MIDINoteDetectorTests: XCTestCase {
         }
 
         for note in noteEvent.notes {
-            let message = MIDIMessage.noteOn(key: note, velocity: 10)
+            let message = MIDIMessage.noteOn(key: UInt8(note), velocity: 10)
             midiNoteDetector.process(midiMessage: message)
         }
 
@@ -70,7 +70,7 @@ class MIDINoteDetectorTests: XCTestCase {
 
         // add expected keys
         for note in noteEvent.notes {
-            midiNoteDetector.process(midiMessage: .noteOn(key: note, velocity: 10))
+            midiNoteDetector.process(midiMessage: .noteOn(key: UInt8(note), velocity: 10))
         }
 
         XCTAssertTrue(notesWereDetected)
